+++ library/src/main/kotlin/tech/coner/crispyfish/StagingRun.kt
@@ -1,0 +1,9 @@
+package tech.coner.crispyfish
+
+import tech.coner.crispyfish.model.Run
+import tech.coner.crispyfish.model.StagingLineRegistration
+
+data class StagingRun(
+    val stagingLineRegistration: StagingLineRegistration?,
+    val run: Run?
+)