+++ library/src/main/kotlin/tech/coner/crispyfish/query/StagingRunsQuery.kt
@@ -1,5 +1,6 @@
 package tech.coner.crispyfish.query
 
+import tech.coner.crispyfish.StagingRun
 import tech.coner.crispyfish.filetype.staging.StagingFile
 import tech.coner.crispyfish.model.Run
 import tech.coner.crispyfish.model.StagingLineRegistration
@@ -8,15 +9,14 @@ class StagingRunsQuery(
     private val stagingFile: StagingFile
 ) {
 
-    fun query(): List<Pair<StagingLineRegistration?, Run?>> {
+    fun query(): List<StagingRun> {
         val reader = stagingFile.modelReader()
-        val pairs = mutableListOf<Pair<StagingLineRegistration?, Run?>>()
-        val lines = stagingFile.file.readLines()
-        for (line in lines) {
-            val registration = reader.readRegistration(line)
-            val run = reader.readRun(line)
-            pairs += registration to run
-        }
-        return pairs
+        return stagingFile.file.readLines()
+            .map { line ->
+                StagingRun(
+                    stagingLineRegistration = reader.readRegistration(line),
+                    run = reader.readRun(line)
+                )
+            }
     }
 }