+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## 0.5.8
+## 0.6.0
 
 - Registration query
   - Includes registrations with incoherent raw/pax results if they are present on disk
   - Includes registrations without category, handicap, or number
+- Staging Runs query
+  - Returns list of StagingRun objects (instead of Kotlin Pairs)
+  - Includes items for every line, including those with fields which couldn't be parsed fully
 
 ## 0.5.7